+++ doc/changelog.md
@@ -33,6 +33,9 @@
 * Handle C++ exceptions without unwinding the whole stack.
   [#99](https://github.com/PennyLaneAI/catalyst/pull/99)
 
+* Support constant negative step sizes in ``@for_loop`` loops.
+  [#129](https://github.com/PennyLaneAI/catalyst/pull/129)
+
 <h3>Breaking changes</h3>
 
 <h3>Bug fixes</h3>